test: split commands tests for windows and unix systems
